requirements.txt
aniso8601==10.0.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
Flask==3.1.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.1.1
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==24.2
pluggy==1.5.0
psycopg2-binary==2.9.9
pytest==8.3.5
python-dotenv==1.1.0
pytz==2024.2
requests==2.32.3
setuptools==70.3.0
six==1.17.0
SQLAlchemy==2.0.29
sqlalchemy-serializer==1.4.22
typing_extensions==4.13.0
urllib3==2.3.0
Werkzeug==3.1.3
WTForms==3.2.1
